requirements.txt
streamlit
openai
feedparser
beautifulsoup4
python-dateutil